Histórico de transação e resolvendo o problema de valor

+++ src/main/java/br/com/projeto/banco_api_simples/Transacao/model/Transacao.java
@@ -21,7 +21,7 @@ public class Transacao {
 
     private String destinatario;
 
-    private int valor;
+    private float  valor;
 
     private LocalDate horarioDaTransacao;
 
@@ -41,10 +41,10 @@ public class Transacao {
         this.destinatario = destinatario;
     }
 
-    public int getValor() {
+    public float  getValor() {
         return valor;
     }
-    public void setValor(int valor) {
+    public void setValor(float  valor) {
         this.valor = valor;
     }
